+++ www/themes/conformance/tailwind.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  content: ['content/**/*.md', 'layouts/**/*.html', '../../content/**/*.md'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}