+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - `--version` flag shows the current version
+- Metadata logging used to associate tests with custom data like versions, specs identifiers, etc.
 
 ## [0.3.0] - 2023-07-31
 ### Added